# backend/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
    "/health": {
      "get": {
        "description": "Inspects live connectivity for PostgreSQL database, Redis caching, and MinIO object storage. Returns 200 OK when operational or degraded, 503 Service Unavailable when critical components fail.",
        "operationId": "HealthController_getHealth",
        "parameters": [],
        "responses": {
          "200": {
            "description": "All services or non-critical services are operational",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HealthResponseDto"
                }
              }
            }
          },
          "503": {
            "description": "One or more critical services (e.g. database) are down",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HealthResponseDto"
                }
              }
            }
          }
        },
        "summary": "Check system health and dependent services status",
        "tags": ["Health & Monitoring"]
      }
    },
    "/auth/login": {
      "post": {
        "description": "Vérifie les identifiants avec Argon2id, gère le verrouillage progressif de compte et émet un Access Token court (15m) ainsi qu'un Refresh Token opaque.",
        "operationId": "AuthController_login",
        "parameters": [
          {
            "name": "user-agent",
            "required": true,
            "in": "header",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LoginDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Authentification réussie",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AuthResponseDto"
                }
              }
            }
          },
          "401": {
            "description": "Identifiants invalides ou compte verrouillé/inactif"
          },
          "429": {
            "description": "Trop de tentatives (Rate limiting)"
          }
        },
        "summary": "Authentification utilisateur avec émission de tokens",
        "tags": ["Authentication & RBAC"]
      }
    },
    "/auth/refresh": {
      "post": {
        "description": "Effectue la rotation du refresh token dans sa famille. Si un token révoqué est réutilisé, toute la famille est invalidée (détection de vol de session).",
        "operationId": "AuthController_refresh",
        "parameters": [
          {
            "name": "user-agent",
            "required": true,
            "in": "header",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RefreshTokenDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Token rafraîchi avec succès",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AuthResponseDto"
                }
              }
            }
          },
          "401": {
            "description": "Token révoqué, expiré ou réutilisation détectée"
          }
        },
        "summary": "Rafraîchissement de tokens avec détection de rejeu",
        "tags": ["Authentication & RBAC"]
      }
    },
    "/auth/logout": {
      "post": {
        "description": "Invalide le refresh token actif et purge le cookie de rafraîchissement.",
        "operationId": "AuthController_logout",
        "parameters": [
          {
            "name": "user-agent",
            "required": true,
            "in": "header",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Déconnexion effectuée",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MessageResponseDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "JWT-auth": []
          }
        ],
        "summary": "Déconnexion et révocation de session",
        "tags": ["Authentication & RBAC"]
      }
    },
    "/auth/me": {
      "get": {
        "description": "Retourne les informations du compte connecté et ses permissions effectives calculées (rôle + overrides individuels).",
        "operationId": "AuthController_getMe",
        "parameters": [],
        "responses": {
          "200": {
            "description": "Profil et permissions",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "JWT-auth": []
          }
        ],
        "summary": "Profil utilisateur et permissions effectives",
        "tags": ["Authentication & RBAC"]
      }
    },
    "/auth/forgot-password": {
      "post": {
        "description": "Génère un token de réinitialisation à usage unique (1h). Réponse générique pour éviter l'énumération d'adresses email.",
        "operationId": "AuthController_forgotPassword",
        "parameters": [
          {
            "name": "user-agent",
            "required": true,
            "in": "header",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ForgotPasswordDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Instructions envoyées",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MessageResponseDto"
                }
              }
            }
          }
        },
        "summary": "Demande de réinitialisation de mot de passe",
        "tags": ["Authentication & RBAC"]
      }
    },
    "/auth/reset-password": {
      "post": {
        "description": "Valide le token, met à jour le hash Argon2id et révoque immédiatement toutes les sessions actives de l'utilisateur.",
        "operationId": "AuthController_resetPassword",
        "parameters": [
          {
            "name": "user-agent",
            "required": true,
            "in": "header",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ResetPasswordDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Mot de passe réinitialisé",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MessageResponseDto"
                }
              }
            }
          }
        },
        "summary": "Réinitialisation de mot de passe via token",
        "tags": ["Authentication & RBAC"]
      }
    },
    "/users": {
      "get": {
        "operationId": "UsersController_getAllUsers",
        "parameters": [],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/UserDto"
                  }
                }
              }
            }
          }
        },
        "security": [
          {
            "JWT-auth": []
          }
        ],
        "summary": "Liste de tous les utilisateurs (Permissions: users:read)",
        "tags": ["User Management"]
      },
      "post": {
        "operationId": "UsersController_createUser",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateUserDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "JWT-auth": []
          }
        ],
        "summary": "Création d'un nouvel utilisateur (Permissions: users:manage)",
        "tags": ["User Management"]
      }
    },
    "/users/{id}": {
      "get": {
        "operationId": "UsersController_getUserById",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "JWT-auth": []
          }
        ],
        "summary": "Détails d'un utilisateur par ID",
        "tags": ["User Management"]
      },
      "put": {
        "operationId": "UsersController_updateUser",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateUserDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "JWT-auth": []
          }
        ],
        "summary": "Mise à jour d'un utilisateur",
        "tags": ["User Management"]
      }
    },
    "/users/{id}/status": {
      "patch": {
        "description": "Si désactivé, toutes les sessions actives sont révoquées immédiatement.",
        "operationId": "UsersController_toggleUserStatus",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "active",
            "required": true,
            "in": "query",
            "schema": {
              "type": "boolean"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "JWT-auth": []
          }
        ],
        "summary": "Activation / Désactivation d'un utilisateur",
        "tags": ["User Management"]
      }
    }
  },
  "info": {
    "title": "HES Logistics API",
    "description": "Enterprise Multi-Agency Logistics, Parcel Transport, Dispatch, Tracking & Billing REST API.",
    "version": "1.0.0",
    "contact": {}
  },
  "tags": [
    {
      "name": "Health & Monitoring",
      "description": "Infrastructure health and readiness probes"
    }
  ],
  "servers": [],
  "components": {
    "securitySchemes": {
      "JWT-auth": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http",
        "description": "Enter your JWT access token"
      }
    },
    "schemas": {
      "ServiceHealthDetailDto": {
        "type": "object",
        "properties": {
          "status": {
            "type": "string",
            "example": "up",
            "enum": ["up", "down"]
          },
          "latencyMs": {
            "type": "number",
            "example": 4,
            "description": "Response latency in milliseconds"
          },
          "error": {
            "type": "string",
            "example": "Connection refused",
            "description": "Error message if service is down"
          }
        },
        "required": ["status", "latencyMs"]
      },
      "HealthServicesMapDto": {
        "type": "object",
        "properties": {
          "database": {
            "$ref": "#/components/schemas/ServiceHealthDetailDto"
          },
          "redis": {
            "$ref": "#/components/schemas/ServiceHealthDetailDto"
          },
          "storage": {
            "$ref": "#/components/schemas/ServiceHealthDetailDto"
          }
        },
        "required": ["database", "redis", "storage"]
      },
      "HealthResponseDto": {
        "type": "object",
        "properties": {
          "status": {
            "type": "string",
            "example": "ok",
            "enum": ["ok", "degraded", "error"]
          },
          "timestamp": {
            "type": "string",
            "example": "2026-08-31T23:00:00.000Z"
          },
          "version": {
            "type": "string",
            "example": "1.0.0"
          },
          "uptimeSeconds": {
            "type": "number",
            "example": 124.5,
            "description": "API uptime in seconds"
          },
          "services": {
            "$ref": "#/components/schemas/HealthServicesMapDto"
          }
        },
        "required": [
          "status",
          "timestamp",
          "version",
          "uptimeSeconds",
          "services"
        ]
      },
      "LoginDto": {
        "type": "object",
        "properties": {
          "email": {
            "type": "string",
            "example": "admin@hes-logistics.com",
            "description": "Email de l'utilisateur"
          },
          "password": {
            "type": "string",
            "example": "AdminSecurePass2026!",
            "description": "Mot de passe de l'utilisateur",
            "minLength": 8
          }
        },
        "required": ["email", "password"]
      },
      "UserDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "example": "45a90d8a-6b80-4966-9eb5-8e7c5eb871b6"
          },
          "email": {
            "type": "string",
            "example": "admin@hes-logistics.com"
          },
          "firstName": {
            "type": "string",
            "example": "Abdou"
          },
          "lastName": {
            "type": "string",
            "example": "Directeur"
          },
          "phone": {
            "type": "object",
            "example": "+33612345678",
            "nullable": true
          },
          "role": {
            "type": "string",
            "example": "SUPER_ADMIN"
          },
          "agencyId": {
            "type": "object",
            "example": "agency-uuid-1234",
            "nullable": true
          },
          "isActive": {
            "type": "boolean",
            "example": true
          },
          "permissions": {
            "example": ["shipments:read", "shipments:create", "users:manage"],
            "type": "array",
            "items": {
              "type": "string"
            }
          }
        },
        "required": [
          "id",
          "email",
          "firstName",
          "lastName",
          "role",
          "isActive",
          "permissions"
        ]
      },
      "TokenPairDto": {
        "type": "object",
        "properties": {
          "accessToken": {
            "type": "string",
            "example": "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9...",
            "description": "JWT Access Token court (15m)"
          },
          "refreshToken": {
            "type": "string",
            "example": "9b1deb4d-3b7d-4bad-9bdd-2b0d7b3dcb6d",
            "description": "Opaque Refresh Token (7j)"
          },
          "expiresIn": {
            "type": "number",
            "example": 900,
            "description": "Durée de validité en secondes (15 min)"
          },
          "tokenType": {
            "type": "string",
            "example": "Bearer"
          }
        },
        "required": ["accessToken", "refreshToken", "expiresIn", "tokenType"]
      },
      "AuthResponseDto": {
        "type": "object",
        "properties": {
          "user": {
            "$ref": "#/components/schemas/UserDto"
          },
          "tokens": {
            "$ref": "#/components/schemas/TokenPairDto"
          }
        },
        "required": ["user", "tokens"]
      },
      "RefreshTokenDto": {
        "type": "object",
        "properties": {
          "refreshToken": {
            "type": "string",
            "example": "9b1deb4d-3b7d-4bad-9bdd-2b0d7b3dcb6d",
            "description": "Refresh Token opaque émis lors du login"
          }
        },
        "required": ["refreshToken"]
      },
      "MessageResponseDto": {
        "type": "object",
        "properties": {
          "message": {
            "type": "string",
            "example": "Opération effectuée avec succès"
          }
        },
        "required": ["message"]
      },
      "ForgotPasswordDto": {
        "type": "object",
        "properties": {
          "email": {
            "type": "string",
            "example": "admin@hes-logistics.com",
            "description": "Email du compte pour réinitialisation du mot de passe"
          }
        },
        "required": ["email"]
      },
      "ResetPasswordDto": {
        "type": "object",
        "properties": {
          "token": {
            "type": "string",
            "example": "reset-token-secret-uuid",
            "description": "Token de réinitialisation sécurisé reçu par email"
          },
          "newPassword": {
            "type": "string",
            "example": "NewSecurePass2026!",
            "description": "Nouveau mot de passe sécurisé",
            "minLength": 8
          }
        },
        "required": ["token", "newPassword"]
      },
      "CreateUserDto": {
        "type": "object",
        "properties": {
          "email": {
            "type": "string",
            "example": "jean.dupont@hes-logistics.com"
          },
          "password": {
            "type": "string",
            "example": "TemporarySecurePassword2026!",
            "minLength": 8
          },
          "firstName": {
            "type": "string",
            "example": "Jean"
          },
          "lastName": {
            "type": "string",
            "example": "Dupont"
          },
          "phone": {
            "type": "string",
            "example": "+33612345678"
          },
          "roleId": {
            "type": "string",
            "example": "role-uuid-1234"
          },
          "agencyId": {
            "type": "string",
            "example": "agency-uuid-5678"
          }
        },
        "required": ["email", "password", "firstName", "lastName", "roleId"]
      },
      "UpdateUserDto": {
        "type": "object",
        "properties": {
          "firstName": {
            "type": "string",
            "example": "Jean"
          },
          "lastName": {
            "type": "string",
            "example": "Dupont"
          },
          "phone": {
            "type": "string",
            "example": "+33612345678"
          },
          "roleId": {
            "type": "string",
            "example": "role-uuid-1234"
          },
          "agencyId": {
            "type": "string",
            "example": "agency-uuid-5678"
          }
        }
      }
    }
  }
}